added typography demo; made mobile changes to left column
