+++ demo/templates/index.html
@@ -1,11 +1,3 @@
 {% extends "app/layout.html" %}
 
-{% block inner_content %}
-  <h1 class="govuk-heading-l">HMLR Design System</h1>
-
-  <ul class="govuk-list govuk-list--bullet">
-    {% for demo in demos %}
-      <li><a href="{{demo}}">{{demo}}</a></li>
-    {% endfor %}
-  </ul>
-{% endblock %}
+{% block inner_content %}{% endblock %}
