+++ hmlr_design_system/components/header-internal/style.scss
