+++ hmlr_design_system/components/links/style.scss
@@ -1,0 +1,12 @@
+.hmlr-link--secondary {
+  @extend %govuk-link;
+  color: $font-color;
+
+  &:visited {
+    color: $font-color;
+  }
+
+  &:hover {
+    color: $font-color--secondary;
+  }
+}
