+++ hmlr_design_system/components/typography/style.scss
@@ -1,0 +1,73 @@
+.hmlr-theme {
+
+  .govuk-heading-xl {
+    font-size: 32px;
+    margin-bottom: 32px;
+  }
+
+  .govuk-heading-l {
+    font-size: 24px;
+    margin-bottom: 24px;
+
+    @include tablet-portrait {
+      font-size: 28px;
+      margin-bottom: 28px;
+    }
+  }
+
+  .govuk-heading-m{
+    font-size: 24px;
+    margin-bottom: 24px;
+  }
+
+  .govuk-heading-s {
+    font-size: 19px;
+    margin-bottom: 19px;
+  }
+
+  .govuk-fieldset__heading {
+    font-size: 28px;
+  }
+
+  .govuk-hint {
+    font-size: $font-size;
+  }
+
+  .hmlr-meta {
+    font-size: $meta-font-size;
+  }
+
+  .govuk-list {
+    font-size: $font-size;
+    list-style-position: inside;
+    margin: 0;
+    padding: 0;
+  }
+
+  .govuk-fieldset__heading {
+    font-size: 28px;
+  }
+
+  .govuk-fieldset__legend--l {
+    font-size: 28px;
+    font-weight: 600;
+  }
+
+  .govuk-fieldset__legend--m {
+    font-size: 24px;
+    font-weight: 600;
+  }
+
+  .govuk-fieldset__legend--s {
+    font-size: 19px;
+    font-weight: 600;
+  }
+
+  .govuk-label {
+    font-size: $font-size;
+  }
+
+  .govuk-error-message {
+    font-size: $font-size;
+  }
+}
